refactor: remove legacy fused moe helper modules
Consolidate the fused MoE helper logic into moe_utils and delete the old comm_utils and experts_selector modules. Update routed experts and quantized MoE paths to use the new layer-owned state, and refresh the affected unit tests accordingly.
